Reorganize storage util
